steady: remove tracer dir
